split out sidebar components
