+++ app/javascript/components/molecules/lease_up_sidebar/LeaseUpStatusButtons.js
@@ -1,0 +1,42 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+import StatusDropdown from '../StatusDropdown'
+
+import { LEASE_UP_STATUS_VALUES } from '../../../utils/statusUtils'
+
+const LeaseUpStatusButtons = ({ status, isLoading, onChangeStatus, onAddCommentClicked }) => (
+  <div className='status-history-buttons'>
+    <StatusDropdown
+      status={status}
+      onChange={onChangeStatus}
+      disabled={isLoading}
+      buttonClasses={['tight-padding']}
+    />
+    <button
+      // ID is just for e2e tests
+      id='add-status-history-comment'
+      className='tertiary tight-padding'
+      disabled={isLoading}
+      onClick={onAddCommentClicked}
+    >
+      Add a comment
+    </button>
+  </div>
+)
+
+LeaseUpStatusButtons.propTypes = {
+  status: PropTypes.oneOf(LEASE_UP_STATUS_VALUES),
+  isLoading: PropTypes.bool,
+  onChangeStatus: PropTypes.func,
+  onAddCommentClicked: PropTypes.func
+}
+
+LeaseUpStatusButtons.defaultProps = {
+  status: null,
+  isLoading: false,
+  onChangeStatus: null,
+  onAddCommentClicked: null
+}
+
+export default LeaseUpStatusButtons
