+++ app/javascript/components/molecules/lease_up_sidebar/StatusItems.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import StatusItem from './StatusItem'
 
-import StatusItemShape from '../../utils/shapes/StatusItemShape'
+import StatusItemShape from '../../../utils/shapes/StatusItemShape'
 
 /**
  * Displays a list of status updates for the Lease-up sidebar.
