+++ app/javascript/components/supplemental_application/SupplementalApplicationContainer.js
@@ -18,7 +18,7 @@ import validate, { touchAllFields } from '~/utils/form/validations'
 import ParkingInformationInputs from './sections/ParkingInformationInputs'
 import { convertPercentAndCurrency } from '../../utils/form/validations'
 import AsymColumnLayout from '../organisms/AsymColumnLayout'
-import LeaseUpSidebar from '../molecules/LeaseUpSidebar'
+import LeaseUpSidebar from '../molecules/lease_up_sidebar/LeaseUpSidebar'
 
 const ConfirmedPreferencesSection = ({
   application,
