+++ app/javascript/packs/pattern_library.js
@@ -28,9 +28,9 @@ import DropdownMenu from 'components/molecules/DropdownMenu'
 import DropdownMenuMultiSelect from 'components/molecules/DropdownMenuMultiSelect'
 import FormGridRow from 'components/molecules/FormGridRow'
 import FormGridTable from 'components/molecules/FormGridTable'
-import LeaseUpSidebar from 'components/molecules/LeaseUpSidebar'
+import LeaseUpSidebar from 'components/molecules/lease_up_sidebar/LeaseUpSidebar'
 import Loading from 'components/molecules/Loading'
-import StatusItems from 'components/molecules/StatusItems'
+import StatusItems from 'components/molecules/lease_up_sidebar/StatusItems'
 import TablePagination from 'components/molecules/TablePagination'
 import TableSimple from 'components/molecules/TableSimple'
 import TabsMenu from 'components/molecules/TabsMenu'
